requirements.txt
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
Jinja2==3.0.1
logger==1.4
loguru==0.5.3
MarkupSafe==2.0.1
numpy==1.20.2
pandas==1.2.4
Pygments==2.8.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
rich==10.1.0
six==1.15.0
typing-extensions==3.7.4.3
win32-setctime==1.0.3
